# api.yaml
swagger: '2.0'
info:
  description: REST API description of the ElasTest TORM Module.
  version: 1.0.0
  title: ElasTest Torm REST API
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8091'
basePath: /v2
...
paths:
...
  /tojob:
    get:
      tags:
        - tojob
      summary: Finds and returns all TOJobs.
      description: Returns all TOJobs for a logged user.
      operationId: getTOJobsUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TOJob'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJobs not found
          schema:
            $ref: '#/definitions/TOJob'
    post:
      tags:
        - tojob
      summary: Creates a new TOJob.
      description: Creates new TOJob with the received description.
      operationId: createTOJobUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: body
          description: TOJjob's desctiption.
          required: true
          schema:
            $ref: '#/definitions/TOJob'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TOJob'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/tojob/{tOJobId}':
    get:
      tags:
        - tojob
      summary: Returns a TOJob.
      description: Returns the TOJob that maches with the given TOJobId.
      operationId: getTOJobByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TOJob'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob not found
          schema:
            $ref: '#/definitions/TOJob'
    put:
      tags:
        - tojob
      summary: Updates the description of a TOJob.
      description: Updates the description of a TOJob.
      operationId: modifyTOJobUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tOJobId
          in: path
          description: Id of TJob that needs to be update.
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          description: TOJob object that needs to be updated.
          required: true
          schema:
            $ref: '#/definitions/TOJob'
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TOJob'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob not exists
          schema:
            $ref: '#/definitions/TOJob'
    delete:
      tags:
        - tojob
      summary: Deletes a TOJob.
      description: Deletes the TOJob that maches with a given TOJobId.
      operationId: deleteTOJobUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob to delete.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob not exists
          schema:
            type: integer
            format: int64
  '/tojob/{tOJobId}/exec':
    get:
      tags:
        - tojob execution
      summary: Returns all executions.
      description: Returns all executions of a TOJobId passed as parameter.
      operationId: getAllTOJobExecutionUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob execution to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/TOJobExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob Executions not found
          schema:
            $ref: '#/definitions/TOJobExecution'
    post:
      tags:
        - tojob execution
      summary: Executes a TOJob.
      description: Executes the TOJob for a given TOJobId.
      operationId: execTOJobUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob to execute.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: integer
            format: int64
        '201':
          description: Created
        '400':
          description: TOJob not found
          schema:
            type: integer
            format: int64
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/tojob/{tOJobId}/exec/{tOJobExecId}':
    get:
      tags:
        - tojob execution
      summary: Returns a TOJobExecution.
      description: 'Returns the TOJobExecution for a given TOJobExecId. '
      operationId: getTOJobExecByIdUsingGET
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob to return.
          required: true
          type: integer
          format: int64
        - name: tOJobExecId
          in: path
          description: ID of TOJob Execution to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/TOJobExecution'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob Execution not found
          schema:
            $ref: '#/definitions/TOJobExecution'
    delete:
      tags:
        - tojob execution
      summary: Deletes a TOJobs Execution.
      description: Deletes a TOJobs Execution for a given TOJobExecId.
      operationId: deleteTOJobExecUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: tOJobId
          in: path
          description: ID of TOJob to return.
          required: true
          type: integer
          format: int64
        - name: tOJobExecId
          in: path
          description: ID of TOJob Execution to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Successful operation
          schema:
            type: integer
            format: int64
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: TOJob Execution not found
          schema:
            type: integer
            format: int64
definitions:
  Project:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      suts:
        type: array
        items:
          $ref: '#/definitions/SutSpecification'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tOJobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tojob:
        type: array
        items:
          $ref: '#/definitions/TOJob'
      tojobs:
        type: array
        items:
          $ref: '#/definitions/TOJob'
  Log:
    type: object
    required:
      - logType
      - logUrl
    properties:
      id:
        type: integer
        format: int64
      logType:
        type: string
        enum:
          - SutLog
          - TestLog
      logUrl:
        type: string
      tJobExec:
        $ref: '#/definitions/TJobExecution'
  SutExecution:
    type: object
    properties:
      deployStatus:
        type: string
        enum:
          - deploying
          - deployed
          - undeploying
          - undeployed
          - error
      id:
        type: integer
        format: int64
        example: 12345678
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
  TJobExecution:
    type: object
    required:
      - duration
      - id
      - result
    properties:
      duration:
        type: integer
        format: int64
      error:
        type: string
      id:
        type: integer
        format: int64
      logs:
        type: array
        description: URLs of logs
        items:
          $ref: '#/definitions/Log'
      result:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      sutExecution:
        $ref: '#/definitions/SutExecution'
      tjob:
        $ref: '#/definitions/TJob'
      tojobExecution:
        $ref: '#/definitions/TOJobExecution'
      tojobExecutionb:
        $ref: '#/definitions/TOJobExecution'
  TOJobExecution:
    type: object
    required:
      - id
      - status
      - tojob
    properties:
      id:
        type: integer
        format: int64
      status:
        type: string
        enum:
          - SUCCESS
          - FAILURE
          - IN PROGRESS
      tJobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tOJob:
        $ref: '#/definitions/TOJob'
      tjobsExecutions:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'
      tojob:
        $ref: '#/definitions/TOJob'
  TOJob:
    type: object
    required:
      - name
      - project
      - toscript
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      project:
        $ref: '#/definitions/Project'
      tOScript:
        type: string
      tojobExecutions:
        type: array
        items:
          $ref: '#/definitions/TOJobExecution'
      toscript:
        type: string
        example: 'tjob1 {sut1, browser}, tjob2 {sut1, browser}, tjob3 {sut2, browser2}'
...
  DeployedSut:
    type: object
    required:
      - url
    properties:
      sshKey:
        type: string
      sutSpecification:
        $ref: '#/definitions/SutSpecification'
      url:
        type: string
  SutSpecification:
    type: object
    required:
      - name
      - project
      - specification
    properties:
      deployedSut:
        type: array
        items:
          $ref: '#/definitions/DeployedSut'
      description:
        type: string
        example: This is a SuT description example
      id:
        type: integer
        format: int64
      name:
        type: string
        example: sut definition 1
      project:
        example: '12345678'
        $ref: '#/definitions/Project'
      specification:
        type: string
      sutExecution:
        type: array
        items:
          $ref: '#/definitions/SutExecution'
      tJobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
      tjobs:
        type: array
        items:
          $ref: '#/definitions/TJob'
...
  TJob:
    type: object
    required:
      - name
    properties:
      id:
        type: integer
        format: int64
      imageName:
        type: string
        example: edujgurjc/torm-test-01
      name:
        type: string
        example: testApp1
      project:
        example: '{ id:"1" }'
        $ref: '#/definitions/Project'
      sut:
        example: '{ id:"1" }'
        $ref: '#/definitions/SutSpecification'
      tjobExecs:
        type: array
        items:
          $ref: '#/definitions/TJobExecution'